requirements.txt
beautifulsoup4==4.6.0

pint
# pyparsing
networkx
watchdog
pyramid
pyramid_jinja2
pyramid_debugtoolbar

nose
reprep
pygments
pygments_markdown_lexer
bcrypt
markdown
junit_xml
lxml
bcrypt
waitress
ruamel.yaml

gitpython
webtest

decorator>=4.1.0
PyContracts==1.8.0
compmake==3.5.20
comptests==1.4.17
DecentLogs==1.1.2
QuickApp==1.3.12
conftools==1.9.1


# -e git+https://github.com/AndreaCensi/compmake.git@mcdp#egg=compmake
# -e git+https://github.com/AndreaCensi/quickapp.git@mcdp#egg=quickapp
# -e git+https://github.com/AndreaCensi/conf_tools.git@mcdp#egg=ConfTools
# -e git+https://github.com/AndreaCensi/comptests.git@mcdp#egg=comptests

#numpy
#matplotlib
chardet
pillow
